requirements.txt
py-cord==2.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
